🐛 fix(AllBugs.tsx): rename UseCaseTable component to AllBugsTable ✨ feat(AllBugs.tsx): add support for fetching bugs data using useBugs hook 🐛 fix(AllBugs.tsx): handle loading and error states in AllBugs component ✨ feat(BugsByState.tsx): add support for fetching bugs by state data using useBugsByState hook 🐛 fix(BugsByState.tsx): handle loading and error states in BugsByState component ✨ feat(BugsByUsecase.tsx): add support for fetching bugs by use case data using useBugsByUseCase hook 🐛 fix(BugsByUsecase.tsx): handle loading and error states in BugsByUsecase component
🔀 refactor(BugsTable/index.tsx): remove unused imports and components, simplify rendering logic
✨ feat(BugsTable/index.tsx): add support for default view based on selected filters
