+++ src/pages/Bugs/Content/BugsTable/BugsByUsecase.tsx
@@ -1,31 +1,38 @@
 import { Accordion, MD } from '@appquality/unguess-design-system';
 import { useMemo } from 'react';
-import { getSelectedFilters } from 'src/features/bugsPage/bugsPageSlice';
 import { useTranslation } from 'react-i18next';
+import { styled } from 'styled-components';
 import { EmptyState } from './components/EmptyState';
-import UseCaseAccordion from './components/SingleGroupAccordion';
-import { BugByUsecaseType } from './types';
 import { CompletionTooltip } from './components/CompletionTooltip';
 import { EmptyGroup } from './components/EmptyGroup';
+import { LoadingState } from './components/LoadingState';
+import { useBugsByUseCase } from './hooks/useBugsByUseCase';
+import BugsByUseCaseAccordion from './components/SingleGroupAccordion';
+
+const Wrapper = styled.div<{
+  isFetching?: boolean;
+}>`
+  padding-top: ${(p) => p.theme.space.lg};
+
+  ${(p) =>
+    p.isFetching &&
+    `
+    opacity: 0.5;
+    pointer-events: none;
+  `}
+`;
 
 export const BugsByUsecase = ({
   campaignId,
-  bugsByUseCases,
+  isDefaultView,
 }: {
   campaignId: number;
-  bugsByUseCases: BugByUsecaseType[];
+  isDefaultView: boolean;
 }) => {
   const { t } = useTranslation();
-  const selectedFilters = getSelectedFilters();
+  const { data, isError, isFetching, isLoading } = useBugsByUseCase(campaignId);
+  const { bugsByUseCases } = data;
 
-  const isDefaultView = useMemo(
-    () =>
-      !selectedFilters.search &&
-      selectedFilters.severities?.length === 0 &&
-      selectedFilters.types?.length === 0 &&
-      !selectedFilters.read,
-    [selectedFilters]
-  );
   const emptyUseCases = useMemo(
     () => bugsByUseCases.filter((item) => item.bugs.length === 0),
     [bugsByUseCases]
@@ -35,59 +42,65 @@ export const BugsByUsecase = ({
     [bugsByUseCases]
   );
 
+  if (isLoading || isError) {
+    return <LoadingState />;
+  }
+
   if (!useCases.length) {
     return <EmptyState />;
   }
 
   return (
-    <Accordion
-      level={3}
-      defaultExpandedSections={Array.from(bugsByUseCases, (_, i) => i)}
-      isExpandable
-      isBare
-    >
-      {useCases.map((item) => (
-        <UseCaseAccordion
-          campaignId={campaignId}
-          key={item.useCase.id}
-          title={
-            <>
-              {item.useCase?.id === -1
-                ? t('__BUGS_PAGE_NO_USECASE', 'Not a specific use case')
-                : item.useCase.title.full}
-              <MD tag="span">{` (${item.bugs.length})`}</MD>
-            </>
-          }
-          item={item}
-          footer={
-            item.useCase?.id !== -1 && (
-              <CompletionTooltip percentage={item.useCase.completion} />
-            )
-          }
-        />
-      ))}
-      {isDefaultView ? (
-        <EmptyGroup isBold>
-          {t(
-            '__BUGS_PAGE_WARNING_POSSIBLE_EMPTY_CASES',
-            "As of now we couldn't find any more bugs in other use cases"
-          )}
-        </EmptyGroup>
-      ) : (
-        <>
-          {emptyUseCases.length > 1 && (
-            <EmptyGroup isBold>
-              {t('__BUGS_PAGE_OTHER_USE_CASES', 'other use cases')}{' '}
-              <MD tag="span">(0)</MD>
-            </EmptyGroup>
-          )}
-          {emptyUseCases.length === 1 && (
-            <EmptyGroup isBold>
-              {emptyUseCases[0].useCase.title.full} <MD tag="span">(0)</MD>
-            </EmptyGroup>
-          )}
-        </>
-      )}
-    </Accordion>
+    <Wrapper isFetching={isFetching}>
+      <Accordion
+        level={3}
+        defaultExpandedSections={Array.from(bugsByUseCases, (_, i) => i)}
+        isExpandable
+        isBare
+      >
+        {useCases.map((item) => (
+          <BugsByUseCaseAccordion
+            campaignId={campaignId}
+            key={item.useCase.id}
+            title={
+              <>
+                {item.useCase?.id === -1
+                  ? t('__BUGS_PAGE_NO_USECASE', 'Not a specific use case')
+                  : item.useCase.title.full}
+                <MD tag="span">{` (${item.bugs.length})`}</MD>
+              </>
+            }
+            item={item}
+            footer={
+              item.useCase?.id !== -1 && (
+                <CompletionTooltip percentage={item.useCase.completion} />
+              )
+            }
+          />
+        ))}
+        {isDefaultView ? (
+          <EmptyGroup isBold>
+            {t(
+              '__BUGS_PAGE_WARNING_POSSIBLE_EMPTY_CASES',
+              "As of now we couldn't find any more bugs in other use cases"
+            )}
+          </EmptyGroup>
+        ) : (
+          <>
+            {emptyUseCases.length > 1 && (
+              <EmptyGroup isBold>
+                {t('__BUGS_PAGE_OTHER_USE_CASES', 'other use cases')}{' '}
+                <MD tag="span">(0)</MD>
+              </EmptyGroup>
+            )}
+            {emptyUseCases.length === 1 && (
+              <EmptyGroup isBold>
+                {emptyUseCases[0].useCase.title.full} <MD tag="span">(0)</MD>
+              </EmptyGroup>
+            )}
+          </>
+        )}
+      </Accordion>
+    </Wrapper>
   );
 };
