+++ src/pages/Bugs/Content/BugsTable/index.tsx
@@ -1,52 +1,32 @@
 import { useAppSelector } from 'src/app/hooks';
-import styled from 'styled-components';
+import { getSelectedFilters } from 'src/features/bugsPage/bugsPageSlice';
+import { useMemo } from 'react';
 import { AllBugs } from './AllBugs';
-import { BugsByState } from './BugsByState';
 import { BugsByUsecase } from './BugsByUsecase';
-import { EmptyState } from './components/EmptyState';
-import { LoadingState } from './components/LoadingState';
-import { useTableData } from './hooks/useTableData';
-
-const Wrapper = styled.div<{
-  isFetching?: boolean;
-}>`
-  padding-top: ${(p) => p.theme.space.lg};
-
-  ${(p) =>
-    p.isFetching &&
-    `
-    opacity: 0.5;
-    pointer-events: none;
-  `}
-`;
+import { BugsByState } from './BugsByState';
 
 const BugsTable = ({ campaignId }: { campaignId: number }) => {
   const { groupBy } = useAppSelector((state) => state.bugsPage);
-  const { data, isLoading, isFetching, isError } = useTableData(campaignId);
-
-  if (isLoading || isError) {
-    return <LoadingState />;
-  }
-
-  if (!data.allBugs.length) {
-    return <EmptyState />;
-  }
+  const selectedFilters = getSelectedFilters();
+  const isDefaultView = useMemo(
+    () =>
+      !selectedFilters.search &&
+      selectedFilters.severities?.length === 0 &&
+      selectedFilters.types?.length === 0 &&
+      !selectedFilters.read,
+    [selectedFilters]
+  );
 
   return (
-    <Wrapper isFetching={isFetching}>
+    <>
       {groupBy === 'usecase' && (
-        <BugsByUsecase
-          campaignId={campaignId}
-          bugsByUseCases={data.bugsByUseCases}
-        />
+        <BugsByUsecase campaignId={campaignId} isDefaultView={isDefaultView} />
       )}
       {groupBy === 'bugState' && (
-        <BugsByState campaignId={campaignId} bugsByStates={data.bugsByStates} />
+        <BugsByState campaignId={campaignId} isDefaultView={isDefaultView} />
       )}
-      {groupBy === 'ungrouped' && (
-        <AllBugs campaignId={campaignId} bugs={data.allBugs} />
-      )}
-    </Wrapper>
+      {groupBy === 'ungrouped' && <AllBugs campaignId={campaignId} />}
+    </>
   );
 };
